backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-multipart
python-dotenv
chromadb
pymongo
